backend/requirements.txt
fastapi==0.115.5
uvicorn==0.32.0
pandas==2.2.3
numpy==2.1.3
python-multipart==0.0.9
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
pydantic==2.9.2
python-dotenv==1.0.1
pdfplumber==0.11.4
